[FEAT] Add prefab Gravity Center
Prefab Gravity Center created, that game objects can attract other game objects with the component "attractable"
