+++ Assets/Prefabs/Gravity Center (central disk).prefab
@@ -1,6 +1,6 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
---- !u!1 &4242702505385478925
+--- !u!1 &4140613173441904769
 GameObject:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -8,39 +8,38 @@ GameObject:
   m_PrefabAsset: {fileID: 0}
   serializedVersion: 6
   m_Component:
-  - component: {fileID: 2106586733613344508}
-  - component: {fileID: 4929280205083729233}
-  - component: {fileID: 4929280205083729232}
-  - component: {fileID: 4929280205083729239}
-  - component: {fileID: 4857353132065613048}
-  m_Layer: 6
-  m_Name: balls
+  - component: {fileID: 1303538527457142971}
+  - component: {fileID: 3084432529721665489}
+  - component: {fileID: 5821710462437223861}
+  - component: {fileID: 1185533612965876740}
+  m_Layer: 0
+  m_Name: Gravity Center (central disk)
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!4 &2106586733613344508
+--- !u!4 &1303538527457142971
 Transform:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 4242702505385478925}
+  m_GameObject: {fileID: 4140613173441904769}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: -0.55, y: 3.32, z: 0}
-  m_LocalScale: {x: 0.25, y: 0.25, z: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1.5, y: 1.5, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!212 &4929280205083729233
+--- !u!212 &3084432529721665489
 SpriteRenderer:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 4242702505385478925}
+  m_GameObject: {fileID: 4140613173441904769}
   m_Enabled: 1
   m_CastShadows: 0
   m_ReceiveShadows: 0
@@ -75,7 +74,7 @@ SpriteRenderer:
   m_SortingLayer: 0
   m_SortingOrder: 0
   m_Sprite: {fileID: -2413806693520163455, guid: a42c3ec54aa81b046bc92beeab0bcba7, type: 3}
-  m_Color: {r: 1, g: 0, b: 0, a: 1}
+  m_Color: {r: 0.5377358, g: 0.5377358, b: 0.5377358, a: 1}
   m_FlipX: 0
   m_FlipY: 0
   m_DrawMode: 0
@@ -85,13 +84,13 @@ SpriteRenderer:
   m_WasSpriteAssigned: 1
   m_MaskInteraction: 0
   m_SpriteSortPoint: 0
---- !u!58 &4929280205083729232
+--- !u!58 &5821710462437223861
 CircleCollider2D:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 4242702505385478925}
+  m_GameObject: {fileID: 4140613173441904769}
   m_Enabled: 1
   m_Density: 1
   m_Material: {fileID: 0}
@@ -100,39 +99,23 @@ CircleCollider2D:
   m_UsedByComposite: 0
   m_Offset: {x: 0, y: 0}
   serializedVersion: 2
-  m_Radius: 0.4988492
---- !u!50 &4929280205083729239
-Rigidbody2D:
-  serializedVersion: 4
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 4242702505385478925}
-  m_BodyType: 0
-  m_Simulated: 1
-  m_UseFullKinematicContacts: 0
-  m_UseAutoMass: 0
-  m_Mass: 1
-  m_LinearDrag: 0
-  m_AngularDrag: 0.05
-  m_GravityScale: 0
-  m_Material: {fileID: 0}
-  m_Interpolate: 0
-  m_SleepingMode: 1
-  m_CollisionDetection: 0
-  m_Constraints: 0
---- !u!114 &4857353132065613048
+  m_Radius: 0.5
+--- !u!114 &1185533612965876740
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 4242702505385478925}
+  m_GameObject: {fileID: 4140613173441904769}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: ca0b74c27d93f7745b9aefc41323db5a, type: 3}
+  m_Script: {fileID: 11500000, guid: 2c30b131efd0e7d42b3e50158f4d0740, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  rotateToCenter: 1
-  currentAttractor: {fileID: 0}
+  attractionLayer:
+    serializedVersion: 2
+    m_Bits: 64
+  gravity: -10
+  effectionRadius: 6
+  AttractedObjects: []
+  planetTransform: {fileID: 0}
