+++ Assets/Scripts/Atractor/Attractor.cs
@@ -1,0 +1,51 @@
+using System.Collections.Generic;
+using System.Linq;
+using UnityEngine;
+
+namespace Atractor
+{
+    public class Attractor : MonoBehaviour
+    {
+        public LayerMask attractionLayer;
+        public float gravity = -10;
+        [SerializeField] private float effectionRadius = 10;
+        public List<Collider2D> AttractedObjects = new List<Collider2D>();
+        [HideInInspector] public Transform planetTransform;
+    
+        void Awake()
+        {
+            planetTransform = GetComponent<Transform>();
+        }
+
+        void Update()
+        {
+            SetAttractedObjects();
+        }
+
+        void FixedUpdate()
+        {
+            AttractObjects();
+        }
+
+        void OnDrawGizmosSelected()
+        {
+            Gizmos.color = Color.yellow;
+            Gizmos.DrawWireSphere(transform.position, effectionRadius);
+        }
+
+        void SetAttractedObjects()
+        {
+            AttractedObjects = Physics2D.OverlapCircleAll(planetTransform.position, effectionRadius, attractionLayer).ToList();// can be optimized
+        }
+
+        void AttractObjects()
+        {
+            for (int i = 0; i < AttractedObjects.Count; i++)
+            {
+                AttractedObjects[i].GetComponent<Attractable.Attractable>().Attract(this);
+            }
+        }
+
+    }
+}
+
